Text parser should be ready for SMTP and POP3
It should be able to correctly parse textual protocols.
The options are quite basic: you can specify the argument separator, whether to remove starting/trailing whitespace characters, prefix or suffix, you can input 2 characters (MINIMUM 2) in the code called "brackets" to parse their content as a single, not splitted piece (e.g. if you specify "<>" as Brackets and the space as separator and you input "<Hello there>" it will interpret it as "Hello there" and not "Hello", "there") and whether to remove or not the brackets from the parsed arguments, all with some minor particularities which will be discussed in the Parser's README or Wiki.