Better configuration, multiline parser fix, USER, PASS, STAT and LIST POP3 commands, mailbox locker
Fixed/improved the configuration parser (this includes a fix to the textual parser for multiline arguments).
Added the following POP3 commands: USER and PASS, for login; STAT, for getting the quantity of messages and their whole in octets; LIST, for listing (or retrieving ID + size) message IDs and sizes.
Added a simple folder locker, to prevent concurrency as the POP3 protocol defines.